pyproject.toml
[project]
name = "alpha"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datamodel-code-generator>=0.26.3",
    "fastapi[standard]>=0.115.4",
    "google-generativeai>=0.8.3",
    "httpx>=0.27.2",
    "pydantic>=2.9.2",
    "rich>=13.9.4",
    "structlog>=24.4.0",
    "anthropic>=0.39.0",
    "pydantic-settings>=2.6.1",
    "pyhumps>=3.8.0",
    "inflect>=5.6.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
]
